refactor: Fix issue with ForumListCreateButton redirect
This commit fixes an issue with the ForumListCreateButton component where the redirect was not working correctly. The setConfirmLoading and hideModal functions were being called before the redirect, causing the modal to close prematurely. By moving the setConfirmLoading and hideModal functions after the redirect, the issue is resolved and the user is properly redirected to the new forum page. This refactor improves the functionality and user experience of the forum list create button.